Shift list and read testing into existing Connect Inject test
