--- acceptance/tests/cli/main_test.go
@@ -1,15 +1,0 @@
-package cli_test
-
-import (
-	"os"
-	"testing"
-
-	testsuite "github.com/hashicorp/consul-k8s/acceptance/framework/suite"
-)
-
-var suite testsuite.Suite
-
-func TestMain(m *testing.M) {
-	suite = testsuite.NewSuite(m)
-	os.Exit(suite.Run())
-}
