+++ acceptance/tests/connect/connect_inject_test.go
@@ -8,6 +8,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/hashicorp/consul-k8s/acceptance/framework/cli"
 	"github.com/hashicorp/consul-k8s/acceptance/framework/consul"
 	"github.com/hashicorp/consul-k8s/acceptance/framework/helpers"
 	"github.com/hashicorp/consul-k8s/acceptance/framework/k8s"
@@ -17,6 +18,8 @@ import (
 	"github.com/stretchr/testify/require"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
+
+const ipv4RegEx = "(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)"
 
 // TestConnectInject tests that Connect works in a default and a secure installation.
 func TestConnectInject(t *testing.T) {
@@ -60,6 +63,9 @@ func TestConnectInject(t *testing.T) {
 
 	for name, c := range cases {
 		t.Run(name, func(t *testing.T) {
+			consulK8sCli, err := cli.NewCLI()
+			require.NoError(t, err)
+
 			cfg := suite.Config()
 			ctx := suite.Environment().DefaultContext(t)
 
@@ -80,6 +86,40 @@ func TestConnectInject(t *testing.T) {
 				connHelper.TestConnectionFailureWithoutIntention(t)
 				connHelper.CreateIntention(t)
 			}
+
+			// Run proxy list and get the two results.
+			listOut, err := consulK8sCli.Run("proxy", "list")
+			require.NoError(t, err)
+			logger.Log(t, string(listOut))
+			list := cli.TranslateListOutput(listOut)
+			require.Equal(t, 2, len(list))
+			for _, proxyType := range list {
+				require.Equal(t, "Sidecar", proxyType)
+			}
+
+			// Run proxy read and check that the connection is present in the output.
+			retrier := &retry.Timer{Timeout: 160 * time.Second, Wait: 2 * time.Second}
+			retry.RunWith(retrier, t, func(r *retry.R) {
+				for podName := range list {
+					out, err := consulK8sCli.Run("proxy", "read", podName)
+					require.NoError(t, err)
+
+					output := string(out)
+					logger.Log(t, output)
+
+					// Both proxies must see their own local agent and app as clusters
+					require.Regexp(r, "local_agent.*STATIC", output)
+					require.Regexp(r, "local_app.*STATIC", output)
+
+					// Static Client must have Static Server as a cluster and endpoint.
+					if strings.Contains(podName, "static-client") {
+						require.Regexp(r, "static-server.*static-server\\.default\\.dc1\\.internal.*EDS", output)
+						require.Regexp(r, ipv4RegEx+".*static-server.default.dc1.internal", output)
+					}
+
+				}
+			})
+
 			connHelper.TestConnectionSuccess(t)
 			connHelper.TestConnectionFailureWhenUnhealthy(t)
 		})
